+++ src/main/java/wooteco/subway/domain/fare/policy/ChildrenDiscountPolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.fare;
+package wooteco.subway.domain.fare.policy;
 
 public class ChildrenDiscountPolicy implements DiscountPolicy {
