+++ src/test/java/wooteco/subway/domain/fare/AgeDiscountPolicyTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.domain.fare.policy.AgeDiscountPolicy;
 import wooteco.subway.domain.fare.vo.Age;
 
 class AgeDiscountPolicyTest {
