+++ src/test/java/wooteco/subway/domain/fare/FreeDiscountPolicyTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.fare.policy.DiscountPolicy;
+import wooteco.subway.domain.fare.policy.FreeDiscountPolicy;
 
 public class FreeDiscountPolicyTest {
