+++ src/test/java/wooteco/subway/domain/fare/TeenagerDiscountPolicyTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.fare.policy.DiscountPolicy;
+import wooteco.subway.domain.fare.policy.TeenagerDiscountPolicy;
 
 public class TeenagerDiscountPolicyTest {
